+++ internal/xsql/badconn/badconn.go
@@ -5,13 +5,20 @@ package badconn
 
 import (
 	"database/sql/driver"
+	"io"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 )
 
 func Map(err error) error {
-	if xerrors.MustDeleteSession(err) {
+	switch {
+	case err == nil:
+		return nil
+	case xerrors.Is(err, io.EOF):
+		return io.EOF
+	case xerrors.Is(err, driver.ErrBadConn), xerrors.MustDeleteSession(err):
 		return driver.ErrBadConn
+	default:
+		return err
 	}
-	return err
 }