+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added grpc and operation codes to errors string description
 * Fixed extended API `internal/xsql/(*conn).GetTables(...)`, `internal/xsql/(*conn).GetAllTables(...)`
 * Fixed test for `database/sql` extended API
 * Extend `scheme.Client` interface with method `Database`